+++ packages/types/eslint.config.js
@@ -1,0 +1,3 @@
+import base from "@friday/config/eslint-base";
+
+export default base;